requirements.txt
Flask==2.0.1
opencv-python==4.5.3.56
numpy==1.21.2
Pillow==8.3.2
OWSLib==0.25.0
torch==1.9.0
torchvision==0.10.0
moviepy==1.0.3
cupy-cuda102 